+++ wp-parsely.php
@@ -28,6 +28,7 @@ namespace Parsely;
 
 use Parsely\Integrations\Amp;
 use Parsely\Integrations\Facebook_Instant_Articles;
+use Parsely\Integrations\Google_Web_Stories;
 use Parsely\Integrations\Integrations;
 use Parsely\UI\Admin_Bar;
 use Parsely\UI\Admin_Warning;
@@ -119,6 +120,7 @@ require __DIR__ . '/src/Integrations/class-integration.php';
 require __DIR__ . '/src/Integrations/class-integrations.php';
 require __DIR__ . '/src/Integrations/class-amp.php';
 require __DIR__ . '/src/Integrations/class-facebook-instant-articles.php';
+require __DIR__ . '/src/Integrations/class-google-web-stories.php';
 
 add_action( 'init', __NAMESPACE__ . '\\parsely_integrations' );
 /**
@@ -132,6 +134,7 @@ function parsely_integrations(): Integrations {
 	$parsely_integrations = new Integrations();
 	$parsely_integrations->register( 'amp', Amp::class );
 	$parsely_integrations->register( 'fbia', Facebook_Instant_Articles::class );
+	$parsely_integrations->register( 'webstories', new Google_Web_Stories( $GLOBALS['parsely'] ) );
 	$parsely_integrations = apply_filters( 'wp_parsely_add_integration', $parsely_integrations );
 	$parsely_integrations->integrate();
